INGM-410: basic example and example with feedbacks configuration.
