+++ examples/commutation_test.py
@@ -1,77 +1,31 @@
+import logging
+import argparse
 from ingeniamotion import MotionController
-from ingeniamotion.enums import SensorType
 
 
-def use_only_incremental_encoders(mc: MotionController):
-    """All feedbacks are using an Incremental encoder.
-
-    Args:
-        mc: Controller to configure the type of encoder for each feedback
-
-    """
-    mc.configuration.set_auxiliar_feedback(SensorType.QEI)
-    mc.configuration.set_commutation_feedback(SensorType.QEI)
-    mc.configuration.set_position_feedback(SensorType.QEI)
-    mc.configuration.set_velocity_feedback(SensorType.QEI)
-    mc.configuration.set_reference_feedback(SensorType.QEI)
+def setup_command():
+    parser = argparse.ArgumentParser(description='Run commutation test')
+    parser.add_argument('dictionary_path', help='path to drive dictionary')
+    parser.add_argument('-ip', default="192.168.2.22", help='drive ip address')
+    parser.add_argument('--axis', default=1, help='drive axis')
+    parser.add_argument('--debug', action='store_true',
+                        help="with this flag test doesn't apply any change")
+    return parser.parse_args()
 
 
-def use_only_absolute_encoders(mc: MotionController):
-    """All feedbacks are using an Absolute encoder.
-
-    Args:
-        mc: Controller to configure the type of encoder for each feedback
-
-    """
-    mc.configuration.set_auxiliar_feedback(SensorType.ABS1)
-    mc.configuration.set_commutation_feedback(SensorType.ABS1)
-    mc.configuration.set_position_feedback(SensorType.ABS1)
-    mc.configuration.set_velocity_feedback(SensorType.ABS1)
-    mc.configuration.set_reference_feedback(SensorType.ABS1)
-
-
-def use_incremental_encoders_and_digital_halls(mc: MotionController):
-    """All feedbacks are using an incremental encoder and a digital halls.
-
-    Args:
-        mc: Controller to configure the type of encoder for each feedback
-
-    """
-    mc.configuration.set_auxiliar_feedback(SensorType.HALLS)
-    mc.configuration.set_commutation_feedback(SensorType.QEI)
-    mc.configuration.set_position_feedback(SensorType.HALLS)
-    mc.configuration.set_velocity_feedback(SensorType.QEI)
-    mc.configuration.set_reference_feedback(SensorType.QEI)
-
-
-def main() -> None:
+def main(args):
+    # Create MotionController instance
     mc = MotionController()
-
-    interface_index = 3
-    slave_id = 1
-    dictionary_path = (
-        "\\\\awe-srv-max-prd\\distext\\products\\CAP-NET\\firmware\\2.5.1\\cap-net-e_eoe_2.5.1.xdf"
-    )
-    mc.communication.connect_servo_ethercat_interface_index(
-        interface_index, slave_id, dictionary_path
-    )
-
-    # Feedbacks configuration:
-    # Indicate the type of encoder you are using for each feedback
-    # There are some examples of feedbacks configuration below.
-    # Uncomment the configuration you want to try:
-    # -------------------------------------------------------------
-    # use_only_incremental_encoders(mc)
-    # use_only_absolute_encoders(mc)
-    use_incremental_encoders_and_digital_halls(mc)
-    # -------------------------------------------------------------
-
+    # Connect Servo with MotionController instance
+    mc.communication.connect_servo_eoe(args.ip, args.dictionary_path)
     # Run Commutation test
-    result = mc.tests.commutation()
-    print(f"Commutation Result: {result['result_message']}")
-
+    result = mc.tests.commutation(axis=args.axis,
+                                  apply_changes=not args.debug)
+    logging.info(result["result_message"])
     mc.communication.disconnect()
 
 
-if __name__ == "__main__":
-    main()
+if __name__ == '__main__':
+    logging.basicConfig(level=logging.DEBUG)
+    args = setup_command()
+    main(args)
